+++ examples/mui/Accordion.variants.tsx
@@ -7,17 +7,14 @@ import Accordion from "@mui/material/Accordion";
 import AccordionDetails from "@mui/material/AccordionDetails";
 import AccordionSummary from "@mui/material/AccordionSummary";
 import Stack from "@mui/material/Stack";
-import Typography from "@mui/material/Typography";
 
 export default () => {
 	return (
-		<Stack spacing={1} direction="column" sx={{ alignSelf: "stretch" }}>
+		<Stack spacing={2} direction="column" sx={{ alignSelf: "stretch" }}>
 			<div>
 				<Accordion variant="elevation">
 					<AccordionSummary>
-						<Typography render={<span />}>
-							<code>elevation</code> variant
-						</Typography>
+						<code>elevation</code> variant
 					</AccordionSummary>
 					<AccordionDetails>
 						An elevation accordion does not include a border around its content
@@ -28,9 +25,7 @@ export default () => {
 			<div>
 				<Accordion variant="outlined">
 					<AccordionSummary>
-						<Typography render={<span />}>
-							<code>outlined</code> variant
-						</Typography>
+						<code>outlined</code> variant
 					</AccordionSummary>
 					<AccordionDetails>
 						An outlined accordion includes a border around its content with