+++ examples/mui/Badge.sizes.tsx
@@ -12,9 +12,8 @@ import svgPlaceholder from "@stratakit/icons/placeholder.svg";
 export default () => {
 	return (
 		<Stack
-			spacing={1}
+			spacing={2}
 			direction="row"
-			useFlexGap
 			sx={{ alignItems: "center", flexWrap: "wrap" }}
 		>
 			<Badge
@@ -25,7 +24,7 @@ export default () => {
 					</>
 				}
 				size="small"
-				inline
+				variant="inline"
 			/>
 			<Badge
 				badgeContent={
@@ -35,7 +34,7 @@ export default () => {
 					</>
 				}
 				size="medium"
-				inline
+				variant="inline"
 			/>
 		</Stack>
 	);