+++ examples/mui/Badge.type.tsx
@@ -8,10 +8,10 @@ import Stack from "@mui/material/Stack";
 
 export default () => {
 	return (
-		<Stack spacing={1} direction="row" useFlexGap sx={{ flexWrap: "wrap" }}>
-			<Badge badgeContent="Strong" type="strong" inline />
-			<Badge badgeContent="Muted" type="muted" inline />
-			<Badge badgeContent="Outlined" type="outlined" inline />
+		<Stack spacing={2} direction="row" sx={{ flexWrap: "wrap" }}>
+			<Badge badgeContent="Strong" type="strong" variant="inline" />
+			<Badge badgeContent="Muted" type="muted" variant="inline" />
+			<Badge badgeContent="Outlined" type="outlined" variant="inline" />
 		</Stack>
 	);
 };