+++ examples/mui/Button._permutations.tsx
@@ -27,18 +27,17 @@ const icons = ["", "end"] as const;
 
 export default () => {
 	return (
-		<Stack spacing={1}>
+		<Stack spacing={2}>
 			{icons.map((icon) => {
 				return (
-					<Stack key={icon} spacing={1}>
+					<Stack key={icon} spacing={2}>
 						{variants.map((variant) => {
 							const variantName = `${variant.charAt(0).toUpperCase()}${variant.slice(1)}`;
 							return (
 								<Stack
 									key={variant}
-									spacing={1}
+									spacing={2}
 									direction="row"
-									useFlexGap
 									sx={{ flexWrap: "wrap" }}
 								>
 									{colors.map((color) => {