+++ packages/bricks/src/Select.tsx
@@ -5,7 +5,8 @@
 
 import * as React from "react";
 import { Role } from "@ariakit/react/role";
-import { forwardRef, isBrowser } from "@stratakit/foundations/secret-internals";
+import { isBrowser } from "@stratakit/internal-utils/dom";
+import { forwardRef } from "@stratakit/internal-utils/react";
 import cx from "classnames";
 import { CaretsUpDown } from "./~utils.icons.js";
 import { useInit } from "./~utils.useInit.js";
@@ -14,7 +15,7 @@ import { useFieldControlType } from "./Field.internal.js";
 import type {
 	BaseProps,
 	FocusableProps,
-} from "@stratakit/foundations/secret-internals";
+} from "@stratakit/internal-utils/props";
 
 const supportsHas = isBrowser && CSS?.supports?.("selector(:has(+ *))");