+++ packages/foundations/src/Root.internal.tsx
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as React from "react";
-import { useIsClient } from "./~hooks.js";
+import { useIsClient } from "@stratakit/internal-utils/hooks";
 
 // ----------------------------------------------------------------------------
 
@@ -26,6 +26,9 @@ interface RootContextValue {
 		rootNode: Document | ShadowRoot,
 		options: { css: string; key: string },
 	) => { cleanup: () => void };
+
+	/** Element to customize the portal provider. */
+	portalProvider?: React.ReactElement | undefined;
 }
 
 // ----------------------------------------------------------------------------
@@ -52,3 +55,5 @@ export const spriteSheetId = "🥝-inline-sprites";
 export const HtmlSanitizerContext = React.createContext<
 	((html: string) => string) | undefined
 >(undefined);
+
+// ----------------------------------------------------------------------------