+++ packages/mui/src/~createTheme.tsx
@@ -10,6 +10,7 @@ import StepConnector from "@mui/material/StepConnector";
 import { createTheme as createMuiTheme } from "@mui/material/styles";
 import cx from "classnames";
 import {
+	MuiAccordionHeadingSlot,
 	MuiAccordionRootSlot,
 	MuiAccordionSummary,
 } from "./~components/MuiAccordion.js";
@@ -35,7 +36,9 @@ import {
 	MuiChipDeleteIcon,
 	MuiChipLabel,
 } from "./~components/MuiChip.js";
+import { MuiDialogPaper } from "./~components/MuiDialog.js";
 import { MuiDivider } from "./~components/MuiDivider.js";
+import { MuiDrawerPaper } from "./~components/MuiDrawer.js";
 import { MuiIconButton } from "./~components/MuiIconButton.js";
 import { MuiInputLabel } from "./~components/MuiInputLabel.js";
 import { MuiMenuListSlot } from "./~components/MuiMenu.js";
@@ -50,14 +53,17 @@ import {
 } from "./~components/MuiTable.js";
 import { MuiTab, MuiTabs } from "./~components/MuiTabs.js";
 import { MuiToggleButton } from "./~components/MuiToggleButton.js";
+import { MuiTooltipPopper } from "./~components/MuiTooltip.js";
 import { MuiTypography, variantMapping } from "./~components/MuiTypography.js";
 import {
+	CalendarIcon,
 	CaretsUpDownIcon,
 	ChevronDownIcon,
 	ChevronLeftDoubleIcon,
 	ChevronLeftIcon,
 	ChevronRightDoubleIcon,
 	ChevronRightIcon,
+	ClockIcon,
 	DismissIcon,
 	ErrorIcon,
 	InfoIcon,
@@ -69,14 +75,8 @@ import type { RoleProps } from "@ariakit/react/role";
 import type { ColorSystemOptions } from "@mui/material/styles";
 import type {} from "@mui/x-date-pickers/themeAugmentation";
 
-interface CreateThemeArgs {
-	portalContainer?: HTMLElement | null;
-}
-
 /** Creates a StrataKit theme for MUI. Should be used with MUI's `ThemeProvider`. */
-function createTheme(args: CreateThemeArgs) {
-	const { portalContainer: container } = args;
-
+function createTheme() {
 	// Map the JS palette back to MUI's own CSS variables, which will then be mapped to the correct StrataKit tokens in CSS.
 	// (This is a fallback for any code that uses MUI's theme.palette values directly instead of CSS variables)
 	const palette = {
@@ -93,9 +93,19 @@ function createTheme(args: CreateThemeArgs) {
 				(shade) => [shade, `var(--stratakit-mui-palette-grey-${shade})`],
 			),
 		),
+
+		tonalOffset: 0.05,
 	} satisfies ColorSystemOptions["palette"];
 
 	return createMuiTheme({
+		spacing: 4,
+		motion: {
+			reducedMotion: "system",
+		},
+		focusVisible: {
+			outline: "var(--🥝focus-outline)",
+			outlineOffset: "var(--🥝focus-outline-offset)",
+		},
 		cssVariables: {
 			nativeColor: true,
 			colorSchemeSelector: "[data-color-scheme='%s']",
@@ -145,6 +155,7 @@ function createTheme(args: CreateThemeArgs) {
 					disableGutters: true,
 					slots: {
 						root: MuiAccordionRootSlot,
+						heading: MuiAccordionHeadingSlot,
 					},
 					slotProps: {
 						region: {
@@ -248,7 +259,12 @@ function createTheme(args: CreateThemeArgs) {
 					slotProps: { badge: { component: MuiBadgeBadge } },
 				},
 			},
-			MuiBottomNavigation: { defaultProps: { component: Role.div } },
+			MuiBottomNavigation: {
+				defaultProps: {
+					component: Role.div,
+					showLabels: true,
+				},
+			},
 			MuiBottomNavigationAction: {
 				defaultProps: {
 					component: MuiBottomNavigationAction,
@@ -276,7 +292,7 @@ function createTheme(args: CreateThemeArgs) {
 			MuiButtonGroup: {
 				defaultProps: {
 					component: Role.div,
-					color: "secondary",
+					color: "secondary" as never,
 					disableRipple: true, // ButtonGroup overrides Button's disableRipple so we need to set it here as well
 				},
 			},
@@ -325,8 +341,54 @@ function createTheme(args: CreateThemeArgs) {
 					},
 				},
 			},
+			MuiCircularProgress: {
+				defaultProps: {
+					enableTrackSlot: true,
+					thickness: 5,
+				},
+			},
 			MuiContainer: { defaultProps: { component: Role.div } },
-			MuiDialog: { defaultProps: { component: Role.div } },
+			MuiDateCalendar: {
+				defaultProps: {
+					slots: {
+						leftArrowIcon: ChevronLeftIcon,
+						rightArrowIcon: ChevronRightIcon,
+						switchViewIcon: ChevronDownIcon,
+					},
+					slotProps: {
+						previousIconButton: { size: "small" },
+						nextIconButton: { size: "small" },
+						day: { nativeButton: true },
+					},
+					dayOfWeekFormatter: (date: Date) =>
+						date
+							.toLocaleDateString(undefined, { weekday: "short" })
+							.slice(0, 2),
+				},
+			},
+			MuiDatePicker: {
+				defaultProps: {
+					slots: {
+						openPickerIcon: withExcludedProps(CalendarIcon, ["ownerState"]),
+					},
+					slotProps: {
+						openPickerButton: {
+							size: "small",
+						},
+					},
+				},
+			},
+			MuiDialog: {
+				defaultProps: {
+					component: Role.div,
+					disableScrollLock: true, // Handled in MuiDialog.css instead.
+					slotProps: {
+						paper: {
+							component: MuiDialogPaper,
+						},
+					},
+				},
+			},
 			MuiDialogContentText: {
 				defaultProps: {
 					component: Role.p,
@@ -340,7 +402,17 @@ function createTheme(args: CreateThemeArgs) {
 				},
 			},
 			MuiDivider: { defaultProps: { component: MuiDivider } },
-			MuiDrawer: { defaultProps: { component: Role.div } },
+			MuiDrawer: {
+				defaultProps: {
+					component: Role.div,
+					disableScrollLock: true, // Handled in MuiDrawer.css instead.
+					slotProps: {
+						paper: {
+							component: MuiDrawerPaper,
+						},
+					},
+				},
+			},
 			MuiFab: {
 				defaultProps: {
 					component: MuiButtonBase,
@@ -364,6 +436,16 @@ function createTheme(args: CreateThemeArgs) {
 			},
 			MuiImageList: { defaultProps: { component: Role.ul } },
 			MuiImageListItem: { defaultProps: { component: Role.li } },
+			MuiInputBase: {
+				defaultProps: {
+					classes: { root: "🥝MuiInput" },
+				},
+			},
+			MuiInput: {
+				defaultProps: {
+					disableUnderline: true,
+				},
+			},
 			MuiInputAdornment: { defaultProps: { component: Role.div } },
 			MuiInputLabel: {
 				defaultProps: {
@@ -389,7 +471,7 @@ function createTheme(args: CreateThemeArgs) {
 					},
 				},
 			},
-			MuiListSubheader: { defaultProps: { component: Role.li } },
+			MuiListSubheader: { defaultProps: { component: Role.div } },
 			MuiMenu: {
 				defaultProps: {
 					component: Role.div,
@@ -406,9 +488,12 @@ function createTheme(args: CreateThemeArgs) {
 			MuiMenuItem: { defaultProps: { component: Role.li } },
 			MuiMenuList: { defaultProps: { component: Role.ul } },
 			MuiMobileStepper: { defaultProps: { component: Role.div } },
-			MuiModal: { defaultProps: { component: Role.div, container } },
+			MuiModal: { defaultProps: { component: Role.div } },
 			MuiOutlinedInput: {
 				defaultProps: {
+					classes: {
+						root: "🥝MuiInput",
+					},
 					notched: false, // Removes masked border from Select
 				},
 			},
@@ -426,23 +511,21 @@ function createTheme(args: CreateThemeArgs) {
 				},
 			},
 			MuiPaper: { defaultProps: { component: Role.div } },
+			MuiPickersInputBase: {
+				defaultProps: {
+					className: "🥝MuiInput",
+				},
+			},
 			MuiPopover: {
 				defaultProps: {
 					component: Role.div,
 					disableScrollLock: true,
-					// Popover passes down `container` prop to `Modal` https://github.com/mui/material-ui/blob/708ef10e874efa63d2e4972bd902befa1912f2dc/packages/mui-material/src/Popover/Popover.js#L389
-					container,
 					slots: {
 						paper: MuiPopoverPaperSlot,
 					},
 					slotProps: {
 						paper: { role: "dialog" },
 					},
-				},
-			},
-			MuiPopper: {
-				defaultProps: {
-					container,
 				},
 			},
 			MuiRadio: {
@@ -486,7 +569,12 @@ function createTheme(args: CreateThemeArgs) {
 				},
 			},
 			MuiSnackbarContent: { defaultProps: { component: Role.div } },
-			MuiStack: { defaultProps: { component: Role.div } },
+			MuiStack: {
+				defaultProps: {
+					component: Role.div,
+					useFlexGap: true,
+				},
+			},
 			MuiStep: { defaultProps: { component: Role.li } },
 			MuiSwitch: { defaultProps: { component: Role.span } },
 			MuiStepper: {
@@ -566,6 +654,18 @@ function createTheme(args: CreateThemeArgs) {
 				},
 			},
 			MuiTextField: { defaultProps: { component: Role.div } },
+			MuiTimePicker: {
+				defaultProps: {
+					slots: {
+						openPickerIcon: withExcludedProps(ClockIcon, ["ownerState"]),
+					},
+					slotProps: {
+						openPickerButton: {
+							size: "small",
+						},
+					},
+				},
+			},
 			MuiToggleButton: { defaultProps: { component: MuiToggleButton } },
 			MuiToolbar: { defaultProps: { component: Role.div } },
 			MuiTooltip: {
@@ -577,11 +677,25 @@ function createTheme(args: CreateThemeArgs) {
 							className: "🥝MuiTooltip",
 						},
 						popper: {
+							component: MuiTooltipPopper,
+							popperOptions: { strategy: "fixed" },
 							modifiers: [
 								{
 									name: "offset",
 									options: {
 										offset: [0, 2],
+									},
+								},
+								{
+									name: "flip",
+									options: {
+										padding: 4,
+									},
+								},
+								{
+									name: "preventOverflow",
+									options: {
+										padding: 4,
 									},
 								},
 							],
@@ -639,6 +753,23 @@ function withRenderProp(
 ) {
 	return React.forwardRef<HTMLDivElement, RoleProps>((props, forwardedRef) => {
 		return <Role render={<DefaultTagName />} {...props} ref={forwardedRef} />;
+	});
+}
+
+// ----------------------------------------------------------------------------
+
+/** HOC that "excludes" certain props from being passed to the specified Component. */
+function withExcludedProps<Element, Props extends object>(
+	Component: React.ComponentType<Props & React.RefAttributes<Element>>,
+	excludedProps: readonly string[],
+) {
+	return React.forwardRef<Element, Props>((props, forwardedRef) => {
+		const filteredProps = { ...props };
+		for (const key of excludedProps) {
+			delete filteredProps[key as keyof React.PropsWithoutRef<Props>];
+		}
+
+		return <Component {...(filteredProps as Props)} ref={forwardedRef} />;
 	});
 }